Pull out TokenManager
